SH-159 Move TcpConnection class to separate file
Created new sub directory ipc and moved new file with TcpConnection class and smart_home_tcp_server files into it. Moved those classes into new namespace SmartHome::IPC.
