+++ central_unit/src/ipc/smart_home_tcp_server.cpp
@@ -1,12 +1,12 @@
+#include "../smart_home_core.h"
 #include "smart_home_tcp_server.h"
-#include  "smart_home_core.h"
+#include "smart_home_tcp_connection.h"
 
 #include <iostream>
 
 namespace bs = boost::system;
 
-namespace SmartHome {
-    // ========================== Tcp Server ==========================
+namespace SmartHome::IPC {
     TcpServer &TcpServer::Instance() {
         static TcpServer ServerInstance;
         return ServerInstance;
@@ -39,7 +39,8 @@ namespace SmartHome {
         mpAcceptor->async_accept(newTcpConnection->getSocket(),
                                  [this, newTcpConnection, ioContext](bs::error_code ec) {
                                      if (!ec && mTcpServerRunning.load() == true) {
-                                         std::cout << "Connection accepted from: " << newTcpConnection->getSocket().remote_endpoint() << std::endl;
+                                         std::cout << "Connection accepted from: " << newTcpConnection->getSocket().
+                                                 remote_endpoint() << std::endl;
                                          newTcpConnection->read();
                                          this->startAsyncAccept(ioContext);
                                      } else if (ec) {
@@ -75,44 +76,4 @@ namespace SmartHome {
     bool TcpServer::isRunning() {
         return mTcpServerRunning.load();
     }
-
-    // ================================================================
-
-    // ======================== Tcp Connection ========================
-    TcpConnection::TcpConnection(ba::io_context &ioContext) : mSocket(ioContext) {
-    };
-
-    TcpConnection::~TcpConnection() {
-        std::cout << "Connection destroyed." << std::endl;
-        if (mSocket.is_open()) {
-            bs::error_code ec;
-            mSocket.close(ec);
-            if (ec) {
-                std::cerr << "Error during connection close: " << ec.message() << std::endl;
-            }
-        }
-    }
-
-    bip::tcp::socket &TcpConnection::getSocket() {
-        return mSocket;
-    }
-
-    void TcpConnection::read() {
-        auto self = shared_from_this();
-
-        ba::async_read_until(mSocket, mStreamBuf, "\n", [this, self](bs::error_code ec, std::size_t bytes_transferred) {
-            if (!ec) {
-                std::istream is(&mStreamBuf);
-                std::string message;
-                std::getline(is, message);
-
-                std::cout << "Received: " << message << std::endl;
-                read();
-            } else {
-                std::cerr << "Error during read: " << ec.message() << std::endl;
-            }
-        });
-    }
-
-    // ================================================================
 }
